requirements.txt
datetime-utils @ git+https://github.com/reimibeta/django-datetime-utils.git
Django==3.1.7
django-admin-list-filter-dropdown==1.0.3
django-cleanup==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-flex-fields==0.9.0
html-render-utils @ git+https://github.com/reimibeta/django-html-render-utils.git
image-utils @ git+https://github.com/reimibeta/django-image-utils.git
Pillow==8.2.0
rest-framework-utils @ git+https://github.com/reimibeta/django-rest-framework-utils.git
# user-models @ git+https://github.com/reimibeta/django-user-models.git